Simple compound (metabolite) databases

CompDb objects provide access to general (metabolite) compound annotations along with metadata information such as the annotation's source, date and release version. The data is stored internally in a database (usually an SQLite database).

CompDb constructs a CompDb object by connecting to the provided database file.

hasMsMsSpectra returns TRUE if MS/MS spectrum data is available in the database and FALSE otherwise.

src_compdb provides access to the CompDb's database via the functionality from the dplyr/dbplyr package.

tables returns a named list (names being table names) with the fields/columns from each table in the database.

dbconn returns the connection (DBIConnection) to the database.

show(object)

CompDb(x, flags = RSQLite::SQLITE_RO)

hasMsMsSpectra(x)

compounds(x, columns, filter, return.type = c("data.frame", "tibble"), ...)

src_compdb(x)

tables(x)

# S4 method for CompDb
dbconn(x)

# S4 method for CompDb
Spectra(object, columns, filter, ...)

# S4 method for CompDb
supportedFilters(object)

Arguments

x
For CompDb: character(1) with the file name of the SQLite compound database. Alternatively it is possible to provide the connection to the database with parameter x.

For all other methods: a `CompDb` object.

flags
flags passed to the SQLite database connection. See SQLite(). Defaults to read-only, i.e. RSQLite::SQLITE_RO.

columns
For compounds, Spectra: character with the names of the database columns that should be retrieved. Use tables() for a list of available column names.

filter
For compounds and Spectra: filter expression or AnnotationFilter() defining a filter to be used to retrieve specific elements from the database.

return.type
For compounds: either "data.frame" or "tibble" to return the result as a data.frame() or tibble(), respectively.

...
additional arguments. Currently not used.

object
For all methods: a CompDb object.

Details

CompDb objects should be created using the constructor function CompDb providing the name of the (SQLite) database file providing the compound annotation data.

Retrieve annotations from the database

Annotations/compound informations can be retrieved from a CompDb database with the compounds and Spectra functions:

compounds extracts compound data from the CompDb object. In contrast to src_compdb it returns the actual data as a data.frame (if return.type = "data.frame") or a tibble::tibble() (if return.type = "tibble"). A compounds call will always return all elements from the compound table (unless a filter is used). Also, the result data.frame will always contain the compound identifier in column "compound_id".

Spectra extract spectra from the database and returns them as a Spectra() object from the Spectra package. Additional annotations requested with the columns parameter are added as additional spectra variables.

Filtering the database

Data access methods such as compounds and Spectra allow to filter the results using specific filter classes and expressions. Filtering uses the concepts from Bioconductor's AnnotationFilter package. All information for a certain compound with the ID "HMDB0000001" can for example be retrieved by passing the filter expression filter = ~ compound_id == "HMDB0000001" to the compounds function.

Use the supportedFilters passing the CompDb object as an argument to get a list of all supported filters.

See also

Author

Johannes Rainer

Examples

## Create a small CompDb from a provided HMDB subset
cmps <- compound_tbl_sdf(system.file("sdf/HMDB_sub.sdf.gz",
    package = "CompoundDb"))

#> Warning: `data_frame()` is deprecated as of tibble 1.1.0.
#> Please use `tibble()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

metad <- data.frame(name = c("source", "url", "source_version",
    "source_date", "organism"),
    value = c("sub_HMDB", "http://www.hmdb.ca", "4", "2017", "Hsapiens"),
    stringsAsFactors = FALSE)

## Load also MS/MS spectra from HMDB xml files
xml_path <- system.file("xml", package = "CompoundDb")
spctra <- msms_spectra_hmdb(xml_path)

#> Going to process 4 xml files.

#> Postprocessing data ... 

#> OK

## Create the SQLite database:
db_file <- createCompDb(cmps, metadata = metad, msms_spectra = spctra,
    path = tempdir())

## Create a CompDb object
cmp_db <- CompDb(db_file)
cmp_db

#> class: CompDb 
#>  data source: sub_HMDB 
#>  version: 4 
#>  organism: Hsapiens 
#>  compound count: 9 
#>  MS/MS spectra count: 4 

## List all tables in the database and their columns
tables(cmp_db)

#> $compound
#> [1] "compound_id" "name"        "inchi"       "inchikey"    "formula"    
#> [6] "exactmass"   "smiles"     
#> 
#> $msms_spectrum
#>  [1] "original_spectrum_id" "compound_id"          "polarity"            
#>  [4] "collision_energy"     "predicted"            "splash"              
#>  [7] "instrument_type"      "instrument"           "precursor_mz"        
#> [10] "spectrum_id"          "msms_mz_range_min"    "msms_mz_range_max"   
#> 
#> $msms_spectrum_peak
#> [1] "spectrum_id" "mz"          "intensity"   "peak_id"    
#> 
#> $synonym
#> [1] "compound_id" "synonym"    
#> 

## Extract a data.frame with the id, name and inchi of all compounds
compounds(cmp_db, columns = c("compound_id", "name", "inchi"))

#>   compound_id                      name
#> 1 HMDB0000001         1-Methylhistidine
#> 2 HMDB0000002        1,3-Diaminopropane
#> 3 HMDB0000005        2-Ketobutyric acid
#> 4 HMDB0000008     2-Hydroxybutyric acid
#> 5 HMDB0000010          2-Methoxyestrone
#> 6 HMDB0000011 (R)-3-Hydroxybutyric acid
#> 7 HMDB0000012              Deoxyuridine
#> 8 HMDB0004370        N-Methyltryptamine
#> 9 HMDB0006719      5,6-trans-Vitamin D3
#>                                                                                                                                                                                             inchi
#> 1                                                                                                         InChI=1S/C7H11N3O2/c1-10-3-5(9-4-10)2-6(8)7(11)12/h3-4,6H,2,8H2,1H3,(H,11,12)/t6-/m0/s1
#> 2                                                                                                                                                              InChI=1S/C3H10N2/c4-2-1-3-5/h1-5H2
#> 3                                                                                                                                                 InChI=1S/C4H6O3/c1-2-3(5)4(6)7/h2H2,1H3,(H,6,7)
#> 4                                                                                                                                            InChI=1S/C4H8O3/c1-2-3(5)4(6)7/h3,5H,2H2,1H3,(H,6,7)
#> 5                                                   InChI=1S/C19H24O3/c1-19-8-7-12-13(15(19)5-6-18(19)21)4-3-11-9-16(20)17(22-2)10-14(11)12/h9-10,12-13,15,20H,3-8H2,1-2H3/t12-,13+,15-,19-/m0/s1
#> 6                                                                                                                                  InChI=1S/C4H8O3/c1-3(5)2-4(6)7/h3,5H,2H2,1H3,(H,6,7)/t3-/m1/s1
#> 7                                                                            InChI=1S/C9H12N2O5/c12-4-6-5(13)3-8(16-6)11-2-1-7(14)10-9(11)15/h1-2,5-6,8,12-13H,3-4H2,(H,10,14,15)/t5-,6+,8+/m0/s1
#> 8                                                                                                                   InChI=1S/C11H14N2/c1-12-7-6-9-8-13-11-5-3-2-4-10(9)11/h2-5,8,12-13H,6-7H2,1H3
#> 9 InChI=1S/C27H44O/c1-19(2)8-6-9-21(4)25-15-16-26-22(10-7-17-27(25,26)5)12-13-23-18-24(28)14-11-20(23)3/h12-13,19,21,24-26,28H,3,6-11,14-18H2,1-2,4-5H3/b22-12+,23-13+/t21-,24+,25-,26?,27-/m1/s1

## Add also the synonyms (aliases) for the compounds. This will cause the
## tables compound and synonym to be joined. The elements of the compound_id
## and name are now no longer unique
res <- compounds(cmp_db, columns = c("compound_id", "name", "synonym"))
head(res)

#>   compound_id              name                synonym
#> 1 HMDB0000001 1-Methylhistidine      1 Methylhistidine
#> 2 HMDB0000001 1-Methylhistidine                 1-MHis
#> 3 HMDB0000001 1-Methylhistidine     1-Methyl histidine
#> 4 HMDB0000001 1-Methylhistidine     1-Methyl-Histidine
#> 5 HMDB0000001 1-Methylhistidine   1-Methyl-L-histidine
#> 6 HMDB0000001 1-Methylhistidine 1-N-Methyl-L-histidine

## Extract spectra for a specific HMDB compound.
sps <- Spectra(cmp_db, filter = ~ compound_id == "HMDB0000001")
sps

#> MSn data (Spectra) with 2 spectra in a MsBackendCompDb backend:
#>     msLevel     rtime scanIndex
#>   <integer> <numeric> <integer>
#> 1        NA        NA        NA
#> 2        NA        NA        NA
#>  ... 27 more variables/columns.
#>  data source: sub_HMDB 
#>  version: 4 
#>  organism: Hsapiens 

## Using return.type = "tibble" the result will be returned as a "tibble"
compounds(cmp_db, return.type = "tibble")

#> # A tibble: 9 x 7
#>   compound_id name     inchi          inchikey   formula exactmass smiles       
#>   <chr>       <chr>    <chr>          <chr>      <chr>       <dbl> <chr>        
#> 1 HMDB0000001 1-Methy… InChI=1S/C7H1… BRMWTNUJH… C7H11N…     169.  "CN1C=NC(C[C…
#> 2 HMDB0000002 1,3-Dia… InChI=1S/C3H1… XFNJVJPLK… C3H10N2      74.1 "NCCCN"      
#> 3 HMDB0000005 2-Ketob… InChI=1S/C4H6… TYEYBOSBB… C4H6O3      102.  "CCC(=O)C(O)…
#> 4 HMDB0000008 2-Hydro… InChI=1S/C4H8… AFENDNXGA… C4H8O3      104.  "CCC(O)C(O)=…
#> 5 HMDB0000010 2-Metho… InChI=1S/C19H… WHEUWNKSC… C19H24…     300.  "[H][C@@]12C…
#> 6 HMDB0000011 (R)-3-H… InChI=1S/C4H8… WHBMMWSBF… C4H8O3      104.  "C[C@@H](O)C…
#> 7 HMDB0000012 Deoxyur… InChI=1S/C9H1… MXHRCPNRJ… C9H12N…     228.  "OC[C@H]1O[C…
#> 8 HMDB0004370 N-Methy… InChI=1S/C11H… NCIKQJBVU… C11H14…     174.  "CNCCC1=CNC2…
#> 9 HMDB0006719 5,6-tra… InChI=1S/C27H… QYSXJUFSX… C27H44O     384.  "CC(C)CCC[C@…

## Use the CompDb in a dplyr setup
library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:S4Vectors':
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from 'package:BiocGenerics':
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

src_cmp <- src_compdb(cmp_db)
src_cmp

#> src:  sqlite 3.33.0 [/tmp/RtmppylJXz/CompDb.Hsapiens.sub_HMDB.4.sqlite]
#> tbls: compound, metadata, msms_spectrum, msms_spectrum_peak, synonym

## Get a tbl for the compound table
cmp_tbl <- tbl(src_cmp, "compound")

## Extract the id, name and inchi
cmp_tbl %>% select(compound_id, name, inchi) %>% collect()

#> # A tibble: 9 x 3
#>   compound_id name              inchi                                           
#>   <chr>       <chr>             <chr>                                           
#> 1 HMDB0000001 1-Methylhistidine InChI=1S/C7H11N3O2/c1-10-3-5(9-4-10)2-6(8)7(11)…
#> 2 HMDB0000002 1,3-Diaminopropa… InChI=1S/C3H10N2/c4-2-1-3-5/h1-5H2              
#> 3 HMDB0000005 2-Ketobutyric ac… InChI=1S/C4H6O3/c1-2-3(5)4(6)7/h2H2,1H3,(H,6,7) 
#> 4 HMDB0000008 2-Hydroxybutyric… InChI=1S/C4H8O3/c1-2-3(5)4(6)7/h3,5H,2H2,1H3,(H…
#> 5 HMDB0000010 2-Methoxyestrone  InChI=1S/C19H24O3/c1-19-8-7-12-13(15(19)5-6-18(…
#> 6 HMDB0000011 (R)-3-Hydroxybut… InChI=1S/C4H8O3/c1-3(5)2-4(6)7/h3,5H,2H2,1H3,(H…
#> 7 HMDB0000012 Deoxyuridine      InChI=1S/C9H12N2O5/c12-4-6-5(13)3-8(16-6)11-2-1…
#> 8 HMDB0004370 N-Methyltryptami… InChI=1S/C11H14N2/c1-12-7-6-9-8-13-11-5-3-2-4-1…
#> 9 HMDB0006719 5,6-trans-Vitami… InChI=1S/C27H44O/c1-19(2)8-6-9-21(4)25-15-16-26…