Annotate m/z values for ion adducts

annotateMz annotates m/z values using either a data.frame containing compound definitions or a CompDb() and a user provided list of potential (expected) adducts. In detail, the function calculates for each input m/z value the masses of the potential ion adducts and compares these with (monoisotopic) masses provided in the reference annotation. A hit is returned if the mass difference is smaller than ppm.

For object being a numeric and compounds a data.frame or equivalend class, the function returns a list of subsets of compounds with rows matching the adducts' mass. If no match is found for the mass of any adducts, a 0-rows data.frame (same columns than compounds is returned).

# S4 method for numeric,DataFrameOrEquivalent
annotateMz(object, compounds, adduct = "[M+H]+", ppm = 10, ...)

# S4 method for DataFrameOrEquivalent,DataFrameOrEquivalent
annotateMz(object, compounds, adduct = "[M+H]+", ppm = 10, mzcol = "mz", ...)

# S4 method for numericOrDataFrameOrEquivalent,CompDb
annotateMz(object, compounds, adduct = "[M+H]+", ppm = 10, ...)

Arguments

object

either a numeric of m/z values or a data.frame (or equivalent) with a column containing the m/z values.

compounds

either a data.frame (or equivalent) or a CompDb() with the reference annotations. A column named "exactmass" is mandatory if compounds is a data.frame.

adduct

adduct definition. A character defining the adduct(s) to look for or a data.frame with adduct definition. See help on mz2mass() from the MetaboCoreUtils package.

ppm

numeric(1) defining the allowed mass difference in ppm.

...

additional parameters for the compounds call if compounds is a CompDb.

mzcol

for object being a data.frame: character(1) defining the column name in object containing the m/z values. Defaults to "mz".

Value

See description above for details.

Author

Johannes Rainer

Examples

## Read compound annotations from a package-internal SDF file containing
## a small subset of HMDB annotations
fl <- system.file("sdf/HMDB_sub.sdf.gz", package = "CompoundDb")
cmps <- compound_tbl_sdf(fl)
cmps
#> # A tibble: 9 × 8
#>   compound_id name    inchi      inchikey  formula exactmass synonyms smiles    
#>   <chr>       <chr>   <chr>      <chr>     <chr>       <dbl> <named > <chr>     
#> 1 HMDB0000001 1-Meth… InChI=1S/… BRMWTNUJ… C7H11N…     169.  <chr [9… "CN1C=NC(…
#> 2 HMDB0000002 1,3-Di… InChI=1S/… XFNJVJPL… C3H10N2      74.1 <chr [8… "NCCCN"   
#> 3 HMDB0000005 2-Keto… InChI=1S/… TYEYBOSB… C4H6O3      102.  <chr [3… "CCC(=O)C…
#> 4 HMDB0000008 2-Hydr… InChI=1S/… AFENDNXG… C4H8O3      104.  <chr [2… "CCC(O)C(…
#> 5 HMDB0000010 2-Meth… InChI=1S/… WHEUWNKS… C19H24…     300.  <chr [7… "[H][C@@]…
#> 6 HMDB0000011 (R)-3-… InChI=1S/… WHBMMWSB… C4H8O3      104.  <chr [2… "C[C@@H](…
#> 7 HMDB0000012 Deoxyu… InChI=1S/… MXHRCPNR… C9H12N…     228.  <chr [1… "OC[C@H]1…
#> 8 HMDB0004370 N-Meth… InChI=1S/… NCIKQJBV… C11H14…     174.  <chr [9… "CNCCC1=C…
#> 9 HMDB0006719 5,6-tr… InChI=1S/… QYSXJUFS… C27H44O     384.  <chr [4… "CC(C)CCC…

## Annotate provided m/z values assuming they represent mass-to-charge
## values for [M+H]+ adducts.
mzs <- c(105.0546, 75.09168, 127.0365, 113.04756, 113.1210)

res <- annotateMz(mzs, cmps, adduct = "[M+H]+")
#> Warning: cannot xtfrm data frames

## res is a `list` with the matching hits for each m/z and the specified
## adduct. If no match was found the result contains only `NA` values.
res
#> $`1`
#> # A tibble: 2 × 10
#>   compound_id name   inchi    inchikey formula exactmass synonyms smiles .adduct
#>   <chr>       <chr>  <chr>    <chr>    <chr>       <dbl> <named > <chr>  <chr>  
#> 1 HMDB0000008 2-Hyd… InChI=1… AFENDNX… C4H8O3       104. <chr [2… CCC(O… [M+H]+ 
#> 2 HMDB0000011 (R)-3… InChI=1… WHBMMWS… C4H8O3       104. <chr [2… C[C@@… [M+H]+ 
#> # … with 1 more variable: .difference <dbl>
#> 
#> $`2`
#> # A tibble: 1 × 10
#>   compound_id name   inchi   inchikey  formula exactmass synonyms smiles .adduct
#>   <chr>       <chr>  <chr>   <chr>     <chr>       <dbl> <named > <chr>  <chr>  
#> 1 HMDB0000002 1,3-D… InChI=… XFNJVJPL… C3H10N2      74.1 <chr [8… NCCCN  [M+H]+ 
#> # … with 1 more variable: .difference <dbl>
#> 
#> $`3`
#> # A tibble: 0 × 10
#> # … with 10 variables: compound_id <chr>, name <chr>, inchi <chr>,
#> #   inchikey <chr>, formula <chr>, exactmass <dbl>, synonyms <named list>,
#> #   smiles <chr>, .adduct <chr>, .difference <dbl>
#> 
#> $`4`
#> # A tibble: 0 × 10
#> # … with 10 variables: compound_id <chr>, name <chr>, inchi <chr>,
#> #   inchikey <chr>, formula <chr>, exactmass <dbl>, synonyms <named list>,
#> #   smiles <chr>, .adduct <chr>, .difference <dbl>
#> 
#> $`5`
#> # A tibble: 0 × 10
#> # … with 10 variables: compound_id <chr>, name <chr>, inchi <chr>,
#> #   inchikey <chr>, formula <chr>, exactmass <dbl>, synonyms <named list>,
#> #   smiles <chr>, .adduct <chr>, .difference <dbl>
#> 

## Using all adducts supported by the `mz2mass` function.
annotateMz(mzs, cmps, ppm = 10, adduct = MetaboCoreUtils::adductNames())
#> Warning: cannot xtfrm data frames
#> $`1`
#> # A tibble: 2 × 10
#>   compound_id name   inchi    inchikey formula exactmass synonyms smiles .adduct
#>   <chr>       <chr>  <chr>    <chr>    <chr>       <dbl> <named > <chr>  <chr>  
#> 1 HMDB0000008 2-Hyd… InChI=1… AFENDNX… C4H8O3       104. <chr [2… CCC(O… [M+H]+ 
#> 2 HMDB0000011 (R)-3… InChI=1… WHBMMWS… C4H8O3       104. <chr [2… C[C@@… [M+H]+ 
#> # … with 1 more variable: .difference <dbl>
#> 
#> $`2`
#> # A tibble: 1 × 10
#>   compound_id name   inchi   inchikey  formula exactmass synonyms smiles .adduct
#>   <chr>       <chr>  <chr>   <chr>     <chr>       <dbl> <named > <chr>  <chr>  
#> 1 HMDB0000002 1,3-D… InChI=… XFNJVJPL… C3H10N2      74.1 <chr [8… NCCCN  [M+H]+ 
#> # … with 1 more variable: .difference <dbl>
#> 
#> $`3`
#> # A tibble: 2 × 10
#>   compound_id name   inchi    inchikey formula exactmass synonyms smiles .adduct
#>   <chr>       <chr>  <chr>    <chr>    <chr>       <dbl> <named > <chr>  <chr>  
#> 1 HMDB0000008 2-Hyd… InChI=1… AFENDNX… C4H8O3       104. <chr [2… CCC(O… [M+Na]+
#> 2 HMDB0000011 (R)-3… InChI=1… WHBMMWS… C4H8O3       104. <chr [2… C[C@@… [M+Na]+
#> # … with 1 more variable: .difference <dbl>
#> 
#> $`4`
#> # A tibble: 1 × 10
#>   compound_id name   inchi   inchikey  formula exactmass synonyms smiles .adduct
#>   <chr>       <chr>  <chr>   <chr>     <chr>       <dbl> <named > <chr>  <chr>  
#> 1 HMDB0000002 1,3-D… InChI=… XFNJVJPL… C3H10N2      74.1 <chr [8… NCCCN  [M+K]+ 
#> # … with 1 more variable: .difference <dbl>
#> 
#> $`5`
#> # A tibble: 0 × 10
#> # … with 10 variables: compound_id <chr>, name <chr>, inchi <chr>,
#> #   inchikey <chr>, formula <chr>, exactmass <dbl>, synonyms <named list>,
#> #   smiles <chr>, .adduct <chr>, .difference <dbl>
#> 

## It is also possible to annotate an input data.frame, in which case each
## input will be duplicated depending on the number of hits for each m/z
mzs_df <- data.frame(id = letters[1:5], mzmed = mzs)

res <- annotateMz(mzs_df, cmps, mzcol = "mzmed", adduct = "[M+H]+")
#> Warning: cannot xtfrm data frames
res
#>     id     mzmed compound_id                      name
#> 1.1  a 105.05460 HMDB0000008     2-Hydroxybutyric acid
#> 1.2  a 105.05460 HMDB0000011 (R)-3-Hydroxybutyric acid
#> 2    b  75.09168 HMDB0000002        1,3-Diaminopropane
#> 3    c 127.03650        <NA>                      <NA>
#> 4    d 113.04756        <NA>                      <NA>
#> 5    e 113.12100        <NA>                      <NA>
#>                                                              inchi
#> 1.1           InChI=1S/C4H8O3/c1-2-3(5)4(6)7/h3,5H,2H2,1H3,(H,6,7)
#> 1.2 InChI=1S/C4H8O3/c1-3(5)2-4(6)7/h3,5H,2H2,1H3,(H,6,7)/t3-/m1/s1
#> 2                               InChI=1S/C3H10N2/c4-2-1-3-5/h1-5H2
#> 3                                                             <NA>
#> 4                                                             <NA>
#> 5                                                             <NA>
#>                        inchikey formula exactmass
#> 1.1 AFENDNXGAFYKQO-UHFFFAOYSA-N  C4H8O3  104.0473
#> 1.2 WHBMMWSBFZVSSR-GSVOUGTGSA-N  C4H8O3  104.0473
#> 2   XFNJVJPLKCPIBV-UHFFFAOYSA-N C3H10N2   74.0844
#> 3                          <NA>    <NA>        NA
#> 4                          <NA>    <NA>        NA
#> 5                          <NA>    <NA>        NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            synonyms
#> 1.1 (RS)-2-Hydroxybutyrate, (RS)-2-Hydroxybutyric acid, 2-Hydroxy-Butanoate, 2-Hydroxy-Butanoic acid, 2-Hydroxy-DL-Butyrate, 2-Hydroxy-DL-Butyric acid, 2-Hydroxy-N-butyrate, 2-Hydroxy-N-butyric acid, 2-Hydroxybutanoate, 2-Hydroxybutanoic acid, 2-Hydroxybutyrate, a-Hydroxy-N-butyrate, a-Hydroxy-N-butyric acid, a-Hydroxybutanoate, a-Hydroxybutanoic acid, a-Hydroxybutyrate, a-Hydroxybutyric acid, alpha-Hydroxy-N-butyrate, alpha-Hydroxy-N-butyric acid, alpha-Hydroxybutanoate, alpha-Hydroxybutanoic acid, alpha-Hydroxybutyrate, alpha-Hydroxybutyric acid, DL-2-Hydroxybutanoate, DL-2-Hydroxybutanoic acid, DL-a-Hydroxybutyrate, DL-a-Hydroxybutyric acid, DL-alpha-Hydroxybutyrate, DL-alpha-Hydroxybutyric acid
#> 1.2                                                                                                                                                                                                     (R)-(-)-b-Hydroxybutyrate, (R)-(-)-b-Hydroxybutyric acid, (R)-(-)-beta-Hydroxybutyrate, (R)-(-)-beta-Hydroxybutyric acid, (R)-3-Hydroxybutanoate, (R)-3-Hydroxybutanoic acid, (R)-3-Hydroxybutyrate, 3-D-Hydroxybutyrate, 3-D-Hydroxybutyric acid, 3-delta-Hydroxybutyrate, 3-delta-Hydroxybutyric acid, BHIB, D-(-)-3-Hydroxybutyrate, D-3-Hydroxybutyrate, D-3-Hydroxybutyric acid, D-beta-Hydroxybutyrate, delta-(-)-3-Hydroxybutyrate, delta-3-Hydroxybutyrate, delta-3-Hydroxybutyric acid, delta-beta-Hydroxybutyrate
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        1,3-Diamino-N-propane, 1,3-Propanediamine, 1,3-Propylenediamine, 1,3-Trimethylenediamine, 3-Aminopropylamine, a,w-Propanediamine, Propane-1,3-diamine, Trimethylenediamine
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                NA
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                NA
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                NA
#>                smiles .adduct .difference
#> 1.1      CCC(O)C(O)=O  [M+H]+  2.0122e-05
#> 1.2 C[C@@H](O)CC(O)=O  [M+H]+  2.0122e-05
#> 2               NCCCN  [M+H]+  5.6700e-06
#> 3                <NA>    <NA>          NA
#> 4                <NA>    <NA>          NA
#> 5                <NA>    <NA>          NA

## Note that `compounds` could also be `CompDb` database, in which a search
## against the annotations provided in that database would be performed.