Filters supported by CompDb

A variety of different filters can be applied to the CompDb object to retrieve only subsets of the data. These filters extend the AnnotationFilter::AnnotationFilter class and support the filtering concepts introduced by Bioconductor's AnnotationFilter package.

The supported filters are:

CompoundIdFilter: filter based on the compound ID.

CompoundNameFilter: filter based on the compound name.

MsmsMzRangeMinFilter: retrieve entries based on the smallest m/z of all peaks of their MS/MS spectra. Requires that MS/MS spectra data are present (i.e. hasMsMsSpectra(cmp_db) returns TRUE).

MsmsMzRangeMaxFilter: retrieve entries based on the largest m/z of all peaks of their MS/MS spectra. Requires that MS/MS spectra data are present (i.e. hasMsMsSpectra(cmp_db) returns TRUE).

CompoundIdFilter(value, condition = "==")

CompoundNameFilter(value, condition = "==")

MsmsMzRangeMinFilter(value, condition = ">=")

MsmsMzRangeMaxFilter(value, condition = "<=")

Arguments

value
The value for the filter. For details see AnnotationFilter::AnnotationFilter().

condition
The condition for the filter. For details see AnnotationFilter::AnnotationFilter().

See also

Author

Johannes Rainer

Examples

library(CompoundDb)

## Create a filter for the compound id
cf <- CompoundIdFilter("comp_a")
cf

#> class: CompoundIdFilter 
#> condition: == 
#> value: comp_a 

## Create a filter using a formula expression
AnnotationFilter(~ compound_id == "comp_b")

#> class: CompoundIdFilter 
#> condition: == 
#> value: comp_b 

## Combine filters
AnnotationFilterList(CompoundIdFilter("a"), CompoundNameFilter("b"))

#> AnnotationFilterList of length 2 
#> compound_id == 'a' & compound_name == 'b'

## Using a formula expression
AnnotationFilter(~ compound_id == "a" | compound_name != "b")

#> AnnotationFilterList of length 2 
#> compound_id == 'a' | compound_name != 'b'