CompDb-based MS spectrum backend

The MsBackendCompDb allows to retrieve MS2 spectra from an CompDb() object/database. The object keeps only a limited amount of data in memory and retrieves the m/z and intensity values from the database on-demand.

It is not intended that users create or use instances of this class directly, the Spectra() call on CompDb() will return a Spectra object that uses this backend.

MsBackendCompDb()

# S4 method for MsBackendCompDb
backendInitialize(object, x, columns, filter, ...)

# S4 method for MsBackendCompDb
show(object)

# S4 method for MsBackendCompDb
as.list(x)

# S4 method for MsBackendCompDb
intensity(object)

# S4 method for MsBackendCompDb
intensity(object) <- value

# S4 method for MsBackendCompDb
mz(object)

# S4 method for MsBackendCompDb
mz(object) <- value

# S4 method for MsBackendCompDb
asDataFrame(object, columns = spectraVariables(object))

# S4 method for MsBackendCompDb
asDataFrame(object) <- value

# S4 method for MsBackendCompDb
$(x, name) <- value

Arguments

object
an MsBackendCompDb instance.

x
an MsBackendCompDb instance.

columns
for asDataFrame: character with names of columns/spectra variables that should be returned. Defaults to spectraVariables(object).

filter
for backendInitialize: optional filter expression to specify which elements to retrieve from the database.

...
ignored.

value
for $<-: the replacement values.

name
for $<-: the name of the spectra variable to replace.

Note

For higher performance it is suggested to change the backend of the Spectra() object to an MsBackendDataFrame() backend with the setBackend() method of Spectra objects.

Methods implemented for MsBackendCompDb

The methods listed here are implemented for the MsBackendCompDb. All other methods are inherited directly from the parent MsBackendDataFrame() class. See that help page for a complete listing of methods.

as.list: gets the full list of peak matrices. Returns a SimpleList(), length equal to the number of spectra and each element being a matrix with columns "mz" and "intensity" with the spectra's m/z and intensity values.

intensity: retrieves the intensity values for all spectra. Returns a NumericList(), each element being the intensity values of one spectrum. The actual intensity values are retrieved from the database.

intensity<-: not supported.

mz: retrieves the m/z values for all spectra. Returns a NumericList(), each element being a numeric with the m/z values of one spectrum. These values are retrieved from the database.

mz<-: not supported.

asDataFrame: returns the complete spectrum data including m/z and intensity values as a DataFrame().

asDataFrame<-: replace the spectrum metadata. Note that columns "mz" and "intensity" are ignored.

$<-: replace or add a spectrum variable. Note that mz, intensity and spectrum_id variables can not be replaced.