CompDb-based MS spectrum backend

The MsBackendCompDb allows to retrieve MS2 spectra from an CompDb() object/database. The object keeps only a limited amount of data in memory and retrieves the m/z and intensity values from the database on-demand.

It is not intended that users create or use instances of this class directly, the Spectra() call on CompDb() will return a Spectra object that uses this backend.

MsBackendCompDb()

# S4 method for MsBackendCompDb
backendInitialize(object, x, columns, filter, ...)

# S4 method for MsBackendCompDb
show(object)

# S4 method for MsBackendCompDb
peaksData(object)

# S4 method for MsBackendCompDb
intensity(object)

# S4 method for MsBackendCompDb
intensity(object) <- value

# S4 method for MsBackendCompDb
mz(object)

# S4 method for MsBackendCompDb
mz(object) <- value

# S4 method for MsBackendCompDb
spectraData(object, columns = spectraVariables(object))

# S4 method for MsBackendCompDb
spectraData(object) <- value

# S4 method for MsBackendCompDb
$(x, name) <- value

Arguments

object

an MsBackendCompDb instance.

x

an MsBackendCompDb instance.

columns

for spectraData: character with names of columns/spectra variables that should be returned. Defaults to spectraVariables(object). Database columns "ms_level", "precursor_mz", "precursor_intensity", "precursor_charge" are mapped to the core Spectra variables msLevel, precursorMz, precursorIntensity and precursorCharge, respectively.

filter

for backendInitialize: optional filter expression to specify which elements to retrieve from the database.

...

ignored.

value

for $<-: the replacement values.

name

for $<-: the name of the spectra variable to replace.

Value

See the description of the respective function.

Note

For higher performance it is suggested to change the backend of the Spectra() object to an MsBackendDataFrame() backend with the setBackend() method of Spectra objects.

Methods implemented for MsBackendCompDb

The methods listed here are implemented for the MsBackendCompDb. All other methods are inherited directly from the parent MsBackendDataFrame() class. See the help of MsBackendDataFrame in the Spectra package for a complete listing of methods.

peaksData: gets the full list of peak matrices. Returns a list(), length equal to the number of spectra and each element being a matrix with columns "mz" and "intensity" with the spectra's m/z and intensity values.

intensity: retrieves the intensity values for all spectra. Returns a NumericList(), each element being the intensity values of one spectrum. The actual intensity values are retrieved from the database.

intensity<-: not supported.

mz: retrieves the m/z values for all spectra. Returns a NumericList(), each element being a numeric with the m/z values of one spectrum. These values are retrieved from the database.

mz<-: not supported.

spectraData: returns the complete spectrum data including m/z and intensity values as a DataFrame().

spectraData<-: replace the spectrum metadata. Note that columns "mz" and "intensity" are ignored.

$<-: replace or add a spectrum variable. Note that mz, intensity and spectrum_id variables can not be replaced.

Author

Johannes Rainer

Examples

## MsBackendCompDb are not expected to be created/instanciated by users
## directly. Users also almost never directly interact with this type of
## object, as it is intented as a pure data backend for the `Spectra` object.
## Users will thus access MS data through such `Spectra` object, which can
## be created for `CompDb` objects using the `Spectra` method (see help
## of the `CompDb` object for more information. This examples shows how
## a `MsBackendCompDb` could be created purely from an SQLite database
## with data from a CompoundDb database.

## Connect to the SQLite database of a `CompDb` distributed via this package
library(RSQLite)
library(Spectra)
cdb <- CompDb(system.file("sql/CompDb.MassBank.sql", package = "CompoundDb"))

be <- backendInitialize(MsBackendCompDb(), cdb)
be
#> MsBackendCompDb with 70 spectra
#>       msLevel     rtime scanIndex
#>     <integer> <numeric> <integer>
#> 1           2        NA        NA
#> 2           2        NA        NA
#> 3           2        NA        NA
#> 4           2        NA        NA
#> 5           2        NA        NA
#> ...       ...       ...       ...
#> 66          2        NA        NA
#> 67          2        NA        NA
#> 68          2        NA        NA
#> 69          2        NA        NA
#> 70          2        NA        NA
#>  ... 38 more variables/columns.
#>  data source: MassBank 
#>  version: 2020.09 
#>  organism: NA 

## Accessing m/z values
mz(be)
#> NumericList of length 70
#> [["1"]] 133.0648 151.0754 155.9743 161.0597 179.0703
#> [["2"]] 133.0648 151.0754 155.9745 161.0597 179.0703
#> [["3"]] 105.0699 133.0648 151.0754 161.0597 179.0703
#> [["4"]] 105.0699 133.0648 151.0754 161.0597 179.0703
#> [["5"]] 105.0699 115.0542 133.0648 151.0754 161.0597 179.0703
#> [["6"]] 506.3324
#> [["7"]] 276.2686 294.2792 312.287 330.2976 470.3112 488.3218 506.3324
#> [["8"]] 56.0502 57.0706 60.045 69.0699 ... 312.287 330.2976 488.3218 506.3324
#> [["9"]] 56.0502 57.0706 60.045 67.0542 ... 242.2115 276.2686 294.2792 312.287
#> [["10"]] 56.0502 57.0706 60.045 67.0542 ... 159.0288 224.2009 276.2686 294.2792
#> ...
#> <60 more elements>