Annotate m/z values for ion adducts

annotateMz annotates m/z values using either a data.frame containing compound definitions or a CompDb() and a user provided list of potential (expected) adducts. In detail, the function calculates for each input m/z value the masses of the potential ion adducts and compares these with (monoisotopic) masses provided in the reference annotation. A hit is returned if the mass difference is smaller than ppm.

For object being a numeric and compounds a data.frame or equivalend class, the function returns a list of subsets of compounds with rows matching the adducts' mass. If no match is found for the mass of any adducts, a row of compounds with all values set to NA is returned. Note that if compounds is [CompDb()], alistofdata.frame`s is returned.s

# S4 method for numeric,DataFrameOrEquivalent
annotateMz(object, compounds, adduct = adducts(), ppm = 10, ...)

# S4 method for DataFrameOrEquivalent,DataFrameOrEquivalent
annotateMz(object, compounds, adduct = adducts(), ppm = 10, mzcol = "mz", ...)

# S4 method for numericOrDataFrameOrEquivalent,CompDb
annotateMz(object, compounds, adduct = adducts(), ppm = 10, ...)

Arguments

object
either a numeric of m/z values or a data.frame (or equivalent) with a column containing the m/z values.

compounds
either a data.frame (or equivalent) or a CompDb() with the reference annotations. A column named "mass" is mandatory if compounds is a data.frame.

adduct
adduct definition. Either a data.frame as returned by adducts() or a character with the names of the adduct(s) in that data.frame.

ppm
numeric(1) defining the allowed mass difference in ppm.

...
additional parameters for the compounds call if compounds is a CompDb.

mzcol
for object being a data.frame: character(1) defining the column name in object containing the m/z values. Defaults to "mz".

Value

See description above for details.

Author

Johannes Rainer

Examples

## Read compound annotations from a package-internal SDF file containing
## a small subset of HMDB annotations
fl <- system.file("sdf/HMDB_sub.sdf.gz", package = "CompoundDb")
cmps <- compound_tbl_sdf(fl)
cmps

#> # A tibble: 9 x 8
#>   compound_id compound_name  inchi    inchi_key  formula  mass synonyms smiles  
#>   <chr>       <chr>          <chr>    <chr>      <chr>   <dbl> <named > <chr>   
#> 1 HMDB0000001 1-Methylhisti… InChI=1… BRMWTNUJH… C7H11N… 169.  <chr [9… "CN1C=N…
#> 2 HMDB0000002 1,3-Diaminopr… InChI=1… XFNJVJPLK… C3H10N2  74.1 <chr [8… "NCCCN" 
#> 3 HMDB0000005 2-Ketobutyric… InChI=1… TYEYBOSBB… C4H6O3  102.  <chr [3… "CCC(=O…
#> 4 HMDB0000008 2-Hydroxybuty… InChI=1… AFENDNXGA… C4H8O3  104.  <chr [2… "CCC(O)…
#> 5 HMDB0000010 2-Methoxyestr… InChI=1… WHEUWNKSC… C19H24… 300.  <chr [7… "[H][C@…
#> 6 HMDB0000011 (R)-3-Hydroxy… InChI=1… WHBMMWSBF… C4H8O3  104.  <chr [2… "C[C@@H…
#> 7 HMDB0000012 Deoxyuridine   InChI=1… MXHRCPNRJ… C9H12N… 228.  <chr [1… "OC[C@H…
#> 8 HMDB0004370 N-Methyltrypt… InChI=1… NCIKQJBVU… C11H14… 174.  <chr [9… "CNCCC1…
#> 9 HMDB0006719 5,6-trans-Vit… InChI=1… QYSXJUFSX… C27H44O 384.  <chr [4… "CC(C)C…

## Annotate provided m/z values assuming they represent mass-to-charge
## values for [M+H]+ adducts.
mzs <- c(105.0546, 75.09168, 127.0365, 113.04756, 113.1210)

res <- annotateMz(mzs, cmps, adduct = "[M+H]+")

## res is a `list` with the matching hits for each m/z and the specified
## adduct. If no match was found the result contains only `NA` values.
res

#> [[1]]
#> # A tibble: 2 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0000008 2-Hydroxybut… InCh… AFENDNXG… C4H8O3   104. <chr [2… CCC(O… [M+H]+
#> 2 HMDB0000011 (R)-3-Hydrox… InCh… WHBMMWSB… C4H8O3   104. <chr [2… C[C@@… [M+H]+
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[2]]
#> # A tibble: 1 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0000002 1,3-Diaminop… InCh… XFNJVJPL… C3H10N2  74.1 <chr [8… NCCCN  [M+H]+
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[3]]
#> # A tibble: 1 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 NA          NA            NA    NA        NA         NA <lgl [1… NA     NA    
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[4]]
#> # A tibble: 1 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 NA          NA            NA    NA        NA         NA <lgl [1… NA     NA    
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[5]]
#> # A tibble: 1 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 NA          NA            NA    NA        NA         NA <lgl [1… NA     NA    
#> # … with 1 more variable: ppm <dbl>
#> 

## Annotating using all adducts defined by `adducts()` and accepting a
## 20ppm difference between the masses
annotateMz(mzs, cmps, ppm = 20)

#> [[1]]
#> # A tibble: 6 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0000008 2-Hydroxybut… InCh… AFENDNXG… C4H8O3   104. <chr [2… CCC(O… [2M+2…
#> 2 HMDB0000011 (R)-3-Hydrox… InCh… WHBMMWSB… C4H8O3   104. <chr [2… C[C@@… [2M+2…
#> 3 HMDB0000008 2-Hydroxybut… InCh… AFENDNXG… C4H8O3   104. <chr [2… CCC(O… [3M+3…
#> 4 HMDB0000011 (R)-3-Hydrox… InCh… WHBMMWSB… C4H8O3   104. <chr [2… C[C@@… [3M+3…
#> 5 HMDB0000008 2-Hydroxybut… InCh… AFENDNXG… C4H8O3   104. <chr [2… CCC(O… [M+H]+
#> 6 HMDB0000011 (R)-3-Hydrox… InCh… WHBMMWSB… C4H8O3   104. <chr [2… C[C@@… [M+H]+
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[2]]
#> # A tibble: 4 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0000002 1,3-Diaminop… InCh… XFNJVJPL… C3H10N2  74.1 <chr [8… NCCCN  [M+H]+
#> 2 HMDB0000002 1,3-Diaminop… InCh… XFNJVJPL… C3H10N2  74.1 <chr [8… NCCCN  [2M+2…
#> 3 HMDB0000002 1,3-Diaminop… InCh… XFNJVJPL… C3H10N2  74.1 <chr [8… NCCCN  [3M+3…
#> 4 HMDB0000001 1-Methylhist… InCh… BRMWTNUJ… C7H11N… 169.  <chr [9… CN1C=… [M+H-…
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[3]]
#> # A tibble: 4 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0000008 2-Hydroxybut… InCh… AFENDNXG… C4H8O3   104. <chr [2… CCC(O… [M+Na…
#> 2 HMDB0000011 (R)-3-Hydrox… InCh… WHBMMWSB… C4H8O3   104. <chr [2… C[C@@… [M+Na…
#> 3 HMDB0000008 2-Hydroxybut… InCh… AFENDNXG… C4H8O3   104. <chr [2… CCC(O… [2M+2…
#> 4 HMDB0000011 (R)-3-Hydrox… InCh… WHBMMWSB… C4H8O3   104. <chr [2… C[C@@… [2M+2…
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[4]]
#> # A tibble: 2 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0000002 1,3-Diaminop… InCh… XFNJVJPL… C3H10N2  74.1 <chr [8… NCCCN  [M+K]+
#> 2 HMDB0000002 1,3-Diaminop… InCh… XFNJVJPL… C3H10N2  74.1 <chr [8… NCCCN  [2M+2…
#> # … with 1 more variable: ppm <dbl>
#> 
#> [[5]]
#> # A tibble: 1 x 10
#>   compound_id compound_name inchi inchi_key formula  mass synonyms smiles adduct
#>   <chr>       <chr>         <chr> <chr>     <chr>   <dbl> <named > <chr>  <chr> 
#> 1 HMDB0004370 N-Methyltryp… InCh… NCIKQJBV… C11H14…  174. <chr [9… CNCCC… [M+H-…
#> # … with 1 more variable: ppm <dbl>
#> 

## It is also possible to annotate an input data.frame, in which case each
## input will be duplicated depending on the number of hits for each m/z
mzs_df <- data.frame(id = letters[1:5], mzmed = mzs)

res <- annotateMz(mzs_df, cmps, mzcol = "mzmed", adduct = "[M+H]+")
res

#>   id     mzmed compound_id             compound_name
#> 1  a 105.05460 HMDB0000008     2-Hydroxybutyric acid
#> 2  a 105.05460 HMDB0000011 (R)-3-Hydroxybutyric acid
#> 3  b  75.09168 HMDB0000002        1,3-Diaminopropane
#> 4  c 127.03650        <NA>                      <NA>
#> 5  d 113.04756        <NA>                      <NA>
#> 6  e 113.12100        <NA>                      <NA>
#>                                                            inchi
#> 1           InChI=1S/C4H8O3/c1-2-3(5)4(6)7/h3,5H,2H2,1H3,(H,6,7)
#> 2 InChI=1S/C4H8O3/c1-3(5)2-4(6)7/h3,5H,2H2,1H3,(H,6,7)/t3-/m1/s1
#> 3                             InChI=1S/C3H10N2/c4-2-1-3-5/h1-5H2
#> 4                                                           <NA>
#> 5                                                           <NA>
#> 6                                                           <NA>
#>                     inchi_key formula     mass
#> 1 AFENDNXGAFYKQO-UHFFFAOYSA-N  C4H8O3 104.0473
#> 2 WHBMMWSBFZVSSR-GSVOUGTGSA-N  C4H8O3 104.0473
#> 3 XFNJVJPLKCPIBV-UHFFFAOYSA-N C3H10N2  74.0844
#> 4                        <NA>    <NA>       NA
#> 5                        <NA>    <NA>       NA
#> 6                        <NA>    <NA>       NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          synonyms
#> 1 (RS)-2-Hydroxybutyrate, (RS)-2-Hydroxybutyric acid, 2-Hydroxy-Butanoate, 2-Hydroxy-Butanoic acid, 2-Hydroxy-DL-Butyrate, 2-Hydroxy-DL-Butyric acid, 2-Hydroxy-N-butyrate, 2-Hydroxy-N-butyric acid, 2-Hydroxybutanoate, 2-Hydroxybutanoic acid, 2-Hydroxybutyrate, a-Hydroxy-N-butyrate, a-Hydroxy-N-butyric acid, a-Hydroxybutanoate, a-Hydroxybutanoic acid, a-Hydroxybutyrate, a-Hydroxybutyric acid, alpha-Hydroxy-N-butyrate, alpha-Hydroxy-N-butyric acid, alpha-Hydroxybutanoate, alpha-Hydroxybutanoic acid, alpha-Hydroxybutyrate, alpha-Hydroxybutyric acid, DL-2-Hydroxybutanoate, DL-2-Hydroxybutanoic acid, DL-a-Hydroxybutyrate, DL-a-Hydroxybutyric acid, DL-alpha-Hydroxybutyrate, DL-alpha-Hydroxybutyric acid
#> 2                                                                                                                                                                                                     (R)-(-)-b-Hydroxybutyrate, (R)-(-)-b-Hydroxybutyric acid, (R)-(-)-beta-Hydroxybutyrate, (R)-(-)-beta-Hydroxybutyric acid, (R)-3-Hydroxybutanoate, (R)-3-Hydroxybutanoic acid, (R)-3-Hydroxybutyrate, 3-D-Hydroxybutyrate, 3-D-Hydroxybutyric acid, 3-delta-Hydroxybutyrate, 3-delta-Hydroxybutyric acid, BHIB, D-(-)-3-Hydroxybutyrate, D-3-Hydroxybutyrate, D-3-Hydroxybutyric acid, D-beta-Hydroxybutyrate, delta-(-)-3-Hydroxybutyrate, delta-3-Hydroxybutyrate, delta-3-Hydroxybutyric acid, delta-beta-Hydroxybutyrate
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      1,3-Diamino-N-propane, 1,3-Propanediamine, 1,3-Propylenediamine, 1,3-Trimethylenediamine, 3-Aminopropylamine, a,w-Propanediamine, Propane-1,3-diamine, Trimethylenediamine
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              NA
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              NA
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              NA
#>              smiles adduct        ppm
#> 1      CCC(O)C(O)=O [M+H]+ 0.19776574
#> 2 C[C@@H](O)CC(O)=O [M+H]+ 0.19776574
#> 3             NCCCN [M+H]+ 0.07039269
#> 4              <NA>   <NA>         NA
#> 5              <NA>   <NA>         NA
#> 6              <NA>   <NA>         NA

## Note that `compounds` could also be `CompDb` database, in which a search
## against the annotations provided in that database would be performed.