Extract compound data from LipidBlast

compound_tbl_lipidblast() extracts basic compound annotations from a LipidBlast file in (json format) downloaded from http://mona.fiehnlab.ucdavis.edu/downloads . Note that no mass spectra data is extracted from the json file.

Usage

compound_tbl_lipidblast(file, collapse = character(), n = -1L, verbose = FALSE)

Arguments

file

character(1) with the name of the file name.

collapse

optional character(1) to be used to collapse multiple values in the columns "synonyms". See examples for details.

n

integer(1) defining the number of rows from the json file that should be read and processed at a time. By default (n = -1L) the complete file is imported and processed. For large json files it is suggested to set e.g. n = 100000 to enable chunk-wise processing and hence reduce the memory demand.

verbose

logical(1) whether some progress information should be provided. Defaults to verbose = FALSE, but for parsing very large files (specifically with chunk-wise processing enabled with n > 0) it might be helpful to set to verbose = TRUE.

Value

A tibble::tibble with general compound information (one row per compound):

compound_id: the ID of the compound.

name: the compound's name.

inchi: the InChI of the compound.

inchikey: the InChI key. NA for all compounds as it is ot provided in MoNA json files.

formula: the chemical formula.

exactmass: the compound's mass.

synonyms: the compound's synonyms (aliases). This type of this column is by default a list to support multiple aliases per compound, unless argument collapse is provided, in which case multiple synonyms are pasted into a single element separated by the value of collapse.

See also

Author

Johannes Rainer and Jan Stanstrup

Examples

## Read compound information from a subset of HMDB
fl <- system.file("json/MoNa-LipidBlast_sub.json", package = "CompoundDb")
cmps <- compound_tbl_lipidblast(fl)
cmps
#> # A tibble: 8 × 7
#>   compound_id      name      inchi           inchikey formula exactmass synonyms
#>   <chr>            <chr>     <chr>           <chr>    <chr>       <dbl> <list>  
#> 1 LipidBlast000001 CerP 24:0 InChI=1S/C24H5… NA       C24H50…      479. <chr>   
#> 2 LipidBlast000002 CerP 26:0 InChI=1S/C26H5… NA       C26H54…      507. <chr>   
#> 3 LipidBlast000003 CerP 28:0 InChI=1S/C28H5… NA       C28H58…      535. <chr>   
#> 4 LipidBlast000004 CerP 30:0 InChI=1S/C30H6… NA       C30H62…      563. <chr>   
#> 5 LipidBlast000005 CerP 32:0 InChI=1S/C32H6… NA       C32H66…      591. <chr>   
#> 6 LipidBlast000006 CerP 32:1 InChI=1S/C32H6… NA       C32H64…      589. <chr>   
#> 7 LipidBlast000007 CerP 34:0 InChI=1S/C34H7… NA       C34H70…      619. <chr>   
#> 8 LipidBlast000008 CerP 36:0 InChI=1S/C36H7… NA       C36H74…      648. <chr>