Extract compound data from LipidBlast

compound_tbl_lipidblast extracts basic comopund annotations from a LipidBlast file in (json format) downloaded from http://mona.fiehnlab.ucdavis.edu/downloads

compound_tbl_lipidblast(file, collapse)

Arguments

file
character(1) with the name of the file name.

collapse
optional character(1) to be used to collapse multiple values in the columns "synonyms". See examples for details.

Value

A tibble::tibble with general compound information (one row per compound):

compound_id: the ID of the compound.

compound_name: the compound's name.

inchi: the InChI of the compound.

inchikey: the InChI key.

formula: the chemical formula.

mass: the compound's mass.

synonyms: the compound's synonyms (aliases). This type of this column is by default a list to support multiple aliases per compound, unless argument collapse is provided, in which case multiple synonyms are pasted into a single element separated by the value of collapse.

See also

Author

Johannes Rainer and Jan Stanstrup

Examples

## Read compound information from a subset of HMDB
fl <- system.file("json/MoNa-LipidBlast_sub.json", package = "CompoundDb")
cmps <- compound_tbl_lipidblast(fl)
cmps

#> # A tibble: 8 x 7
#>   compound_id  compound_name inchi            inchi_key formula   mass synonyms 
#>   <chr>        <chr>         <chr>            <chr>     <chr>    <dbl> <chr>    
#> 1 LipidBlast0… CerP 24:0     InChI=1S/C24H50… NA        C24H50N…  479. CerP(d14…
#> 2 LipidBlast0… CerP 26:0     InChI=1S/C26H54… NA        C26H54N…  507. CerP(d14…
#> 3 LipidBlast0… CerP 28:0     InChI=1S/C28H58… NA        C28H58N…  535. CerP(d14…
#> 4 LipidBlast0… CerP 30:0     InChI=1S/C30H62… NA        C30H62N…  563. CerP(d14…
#> 5 LipidBlast0… CerP 32:0     InChI=1S/C32H66… NA        C32H66N…  591. CerP(d14…
#> 6 LipidBlast0… CerP 32:1     InChI=1S/C32H64… NA        C32H64N…  589. CerP(d14…
#> 7 LipidBlast0… CerP 34:0     InChI=1S/C34H70… NA        C34H70N…  619. CerP(d14…
#> 8 LipidBlast0… CerP 36:0     InChI=1S/C36H74… NA        C36H74N…  648. CerP(d14…