Import MS/MS spectra from MoNa

msms_spectra_mona imports MS/MS spectra from a MoNa (Massbank of North America, http://mona.fiehnlab.ucdavis.edu/downloads) SDF file and returns the data as a data.frame.

Depending on the parameter collapsed, the returned data.frame is either collapsed, meaning that each row represents data from one spectrum, or expanded with one row for each m/z and intensity pair for each spectrum. Columns "mz" and "intensity" are of type list for collapsed = TRUE and numeric for collapsed = FALSE.

msms_spectra_mona(x, collapsed = TRUE)

Arguments

x
character(1): with the path to directory containing the xml files.

collapsed
logical(1) whether the returned data.frame should be collapsed or expanded. See description for more details.

Value

data.frame with as many rows as there are peaks and columns:

spectrum_id (integer): an arbitrary, unique ID for each spectrum.

original_spectrum_id (character): The ID from the spectrum as specified in the MoNa SDF.

compound_id (character): the compound ID the spectrum is associated with.

polarity (integer): 0 for negative, 1 for positive, NA for not set.

collision_energy (character): collision energy voltage.

predicted (logical): whether the spectrum is predicted or experimentally verified.

splash (character): NA since SPLASH (SPectraL hASH) keys are not provided.

instrument_type (character): the type of MS instrument on which the spectrum was measured.

instrument (character): the MS instrument.

precursor_mz (numeric): precursor m/z.

mz (numeric or list of numeric): m/z values of the spectrum.

intensity (numeric or list of numeric): intensity of the spectrum.

Note

The identifiers provided by MoNa are used as spectrum_id. Note also that the MoNa data is not normalized in the sense that each spectrum is associated to one compound and the compound data is partially redundant. Also, MoNa does not provide a splash for a spectrum, hence the corresponding column will only contain NA.

See also

Author

Johannes Rainer

Examples

## Define the test file containing the data
fl <- system.file("sdf/MoNa_export-All_Spectra_sub.sdf.gz",
    package = "CompoundDb")
## Import spectrum data from the SDF file with a subset of the MoNa data
msms_spectra_mona(fl)

#> Reading SDF file ... 

#> OK

#> Extracting 7 spectra ... 

#> OK

#>      original_spectrum_id compound_id polarity  collision_energy predicted
#> CMP1             AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> CMP2             AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> CMP3             AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> CMP4             AU100802        CMP4        1             20 eV        NA
#> CMP5             AU100803        CMP5        1             30 eV        NA
#> CMP6             AU100804        CMP6        1             40 eV        NA
#> CMP7             AU100805        CMP7        1             50 eV        NA
#>      splash instrument_type          instrument precursor_mz spectrum_id
#> CMP1   <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> CMP2   <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> CMP3   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> CMP4   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> CMP5   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> CMP6   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> CMP7   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                              mz
#> CMP1                                                                                                                                                                                           53.0389, 54.0333, 55.0178, 60.0552, 65.0382, 66.0423, 68.0490, 78.0333, 79.0177, 92.0498, 93.0532, 96.0443, 108.0457, 109.0483, 110.0609, 120.0562, 130.0172, 132.0138, 156.0118, 157.0150, 158.0080, 174.0228, 184.0757, 191.9647, 219.0438, 285.0221, 287.0184
#> CMP2                                                                                                                                                                                           53.0389, 54.0333, 55.0178, 60.0552, 65.0382, 66.0423, 68.0490, 78.0333, 79.0177, 92.0498, 93.0532, 96.0443, 108.0457, 109.0483, 110.0609, 120.0562, 130.0172, 132.0138, 156.0118, 157.0150, 158.0080, 174.0228, 184.0757, 191.9647, 219.0438, 285.0221, 287.0184
#> CMP3 53.0379, 54.0335, 55.0176, 65.0379, 68.0491, 69.0329, 78.0334, 79.0178, 80.0489, 81.0444, 82.0284, 92.0499, 93.0558, 94.0647, 95.0608, 96.0443, 108.0461, 109.0497, 110.0616, 111.0651, 120.0565, 122.0716, 123.0794, 124.0872, 125.0905, 126.0663, 127.0697, 156.0117, 157.0148, 158.0072, 174.0224, 186.0334, 187.0368, 188.0128, 188.0291, 204.0445, 205.0473, 206.0406, 213.1141, 214.1167, 215.0927, 215.1291, 279.0925, 280.0953, 281.0725, 282.0742
#> CMP4                                                                                                                                                                                                                                                                                             122.0703, 124.0861, 125.0892, 149.0227, 156.0104, 157.0129, 158.0061, 174.0209, 186.0321, 187.0346, 188.0285, 204.0431, 213.1128, 214.1159, 215.1281, 279.0909
#> CMP5                                                                                                                                                                                                                                                                                   108.0441, 122.0704, 123.0781, 124.0861, 125.0889, 149.0221, 156.0106, 158.0064, 186.0323, 187.0355, 196.0858, 204.0429, 205.0455, 213.1123, 214.1155, 215.1283, 279.0903
#> CMP6                                                                                                                                                                                                                                 108.0445, 122.0702, 123.0772, 124.0862, 125.0890, 134.0701, 149.0224, 154.0624, 155.0685, 156.0104, 157.0126, 172.0852, 186.0324, 196.0852, 197.0903, 198.0888, 204.0427, 205.0463, 206.0375, 212.1036, 213.1121, 214.1152
#> CMP7                                                                                                                                                                                                             108.0453, 122.0703, 123.0780, 124.0859, 125.0891, 149.0231, 154.0639, 155.0605, 155.0714, 156.0100, 169.0745, 171.0781, 172.0869, 181.0634, 186.1022, 195.0786, 196.0859, 197.0856, 198.0886, 199.0904, 204.0438, 212.1048, 213.1122, 214.1153
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  intensity
#> CMP1                                                                                                                                                                                                       0.594951, 0.566811, 0.522592, 0.542692, 3.822962, 0.506512, 7.963499, 0.727609, 1.057244, 7.702203, 0.731629, 0.623091, 12.172375, 1.181862, 4.904325, 3.095353, 5.656054, 1.515517, 100.000000, 8.884065, 3.891301, 0.751729, 0.619071, 0.590931, 0.723589, 3.694324, 0.840167
#> CMP2                                                                                                                                                                                                       0.594951, 0.566811, 0.522592, 0.542692, 3.822962, 0.506512, 7.963499, 0.727609, 1.057244, 7.702203, 0.731629, 0.623091, 12.172375, 1.181862, 4.904325, 3.095353, 5.656054, 1.515517, 100.000000, 8.884065, 3.891301, 0.751729, 0.619071, 0.590931, 0.723589, 3.694324, 0.840167
#> CMP3 0.894101, 0.661867, 0.598003, 8.717487, 13.013818, 1.640153, 1.477589, 2.261379, 1.431143, 1.950766, 0.606712, 30.585230, 2.844868, 1.686600, 3.027752, 1.300511, 33.946818, 2.360079, 6.107757, 0.519624, 1.962378, 6.078727, 2.246865, 71.211681, 6.398049, 17.911054, 0.595100, 82.855318, 5.739085, 1.544357, 1.106015, 11.263353, 0.775081, 1.637250, 0.534138, 100.000000, 6.972829, 3.358686, 18.259405, 2.241059, 3.071296, 1.320831, 61.483976, 8.438806, 7.837901, 1.222132
#> CMP4                                                                                                                                                                                                                                                                                                                  0.766124, 36.693459, 1.930893, 0.828453, 53.249536, 2.999571, 1.778967, 0.627183, 22.621444, 1.719235, 0.646661, 100.000000, 8.749399, 1.407591, 0.658348, 80.894937
#> CMP5                                                                                                                                                                                                                                                                                                        1.285794, 6.630847, 2.170942, 100.000000, 6.093546, 1.388285, 50.043481, 1.615007, 15.118951, 1.323064, 1.220573, 70.964035, 4.931983, 22.610100, 3.003292, 0.804398, 3.580968
#> CMP6                                                                                                                                                                                                                                                       1.153673, 5.323878, 2.202467, 100.000000, 6.847126, 0.714179, 1.747990, 0.644259, 0.624282, 10.373071, 0.933926, 0.564351, 3.845578, 5.209010, 1.378415, 2.362283, 15.422264, 0.869001, 0.759127, 0.659242, 18.109174, 2.577036
#> CMP7                                                                                                                                                                                                                                      1.770916, 2.803951, 2.792598, 100.000000, 7.901010, 1.623340, 2.111477, 2.463390, 2.690430, 2.713134, 1.475763, 1.555228, 1.271427, 0.930866, 1.033034, 1.555228, 7.628562, 3.871041, 5.903054, 0.998978, 2.622318, 2.327165, 9.342718, 1.725508

## Import the data as an *expanded* data frame, i.e. with a row for each
## single m/z (intensity) value.
msms_spectra_mona(fl, collapsed = FALSE)

#> Reading SDF file ... 

#> OK

#> Extracting 7 spectra ... 

#> OK

#>     original_spectrum_id compound_id polarity  collision_energy predicted
#> 1               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 2               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 3               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 4               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 5               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 6               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 7               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 8               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 9               AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 10              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 11              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 12              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 13              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 14              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 15              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 16              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 17              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 18              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 19              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 20              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 21              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 22              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 23              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 24              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 25              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 26              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 27              AU100601        CMP1        1 Ramp 21.1-31.6 eV        NA
#> 28              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 29              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 30              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 31              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 32              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 33              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 34              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 35              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 36              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 37              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 38              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 39              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 40              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 41              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 42              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 43              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 44              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 45              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 46              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 47              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 48              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 49              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 50              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 51              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 52              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 53              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 54              AU100701        CMP2        1 Ramp 21.1-31.6 eV        NA
#> 55              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 56              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 57              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 58              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 59              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 60              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 61              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 62              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 63              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 64              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 65              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 66              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 67              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 68              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 69              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 70              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 71              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 72              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 73              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 74              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 75              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 76              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 77              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 78              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 79              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 80              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 81              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 82              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 83              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 84              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 85              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 86              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 87              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 88              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 89              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 90              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 91              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 92              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 93              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 94              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 95              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 96              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 97              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 98              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 99              AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 100             AU100801        CMP3        1 Ramp 20.8-31.3 eV        NA
#> 101             AU100802        CMP4        1             20 eV        NA
#> 102             AU100802        CMP4        1             20 eV        NA
#> 103             AU100802        CMP4        1             20 eV        NA
#> 104             AU100802        CMP4        1             20 eV        NA
#> 105             AU100802        CMP4        1             20 eV        NA
#> 106             AU100802        CMP4        1             20 eV        NA
#> 107             AU100802        CMP4        1             20 eV        NA
#> 108             AU100802        CMP4        1             20 eV        NA
#> 109             AU100802        CMP4        1             20 eV        NA
#> 110             AU100802        CMP4        1             20 eV        NA
#> 111             AU100802        CMP4        1             20 eV        NA
#> 112             AU100802        CMP4        1             20 eV        NA
#> 113             AU100802        CMP4        1             20 eV        NA
#> 114             AU100802        CMP4        1             20 eV        NA
#> 115             AU100802        CMP4        1             20 eV        NA
#> 116             AU100802        CMP4        1             20 eV        NA
#> 117             AU100803        CMP5        1             30 eV        NA
#> 118             AU100803        CMP5        1             30 eV        NA
#> 119             AU100803        CMP5        1             30 eV        NA
#> 120             AU100803        CMP5        1             30 eV        NA
#> 121             AU100803        CMP5        1             30 eV        NA
#> 122             AU100803        CMP5        1             30 eV        NA
#> 123             AU100803        CMP5        1             30 eV        NA
#> 124             AU100803        CMP5        1             30 eV        NA
#> 125             AU100803        CMP5        1             30 eV        NA
#> 126             AU100803        CMP5        1             30 eV        NA
#> 127             AU100803        CMP5        1             30 eV        NA
#> 128             AU100803        CMP5        1             30 eV        NA
#> 129             AU100803        CMP5        1             30 eV        NA
#> 130             AU100803        CMP5        1             30 eV        NA
#> 131             AU100803        CMP5        1             30 eV        NA
#> 132             AU100803        CMP5        1             30 eV        NA
#> 133             AU100803        CMP5        1             30 eV        NA
#> 134             AU100804        CMP6        1             40 eV        NA
#> 135             AU100804        CMP6        1             40 eV        NA
#> 136             AU100804        CMP6        1             40 eV        NA
#> 137             AU100804        CMP6        1             40 eV        NA
#> 138             AU100804        CMP6        1             40 eV        NA
#> 139             AU100804        CMP6        1             40 eV        NA
#> 140             AU100804        CMP6        1             40 eV        NA
#> 141             AU100804        CMP6        1             40 eV        NA
#> 142             AU100804        CMP6        1             40 eV        NA
#> 143             AU100804        CMP6        1             40 eV        NA
#> 144             AU100804        CMP6        1             40 eV        NA
#> 145             AU100804        CMP6        1             40 eV        NA
#> 146             AU100804        CMP6        1             40 eV        NA
#> 147             AU100804        CMP6        1             40 eV        NA
#> 148             AU100804        CMP6        1             40 eV        NA
#> 149             AU100804        CMP6        1             40 eV        NA
#> 150             AU100804        CMP6        1             40 eV        NA
#> 151             AU100804        CMP6        1             40 eV        NA
#> 152             AU100804        CMP6        1             40 eV        NA
#> 153             AU100804        CMP6        1             40 eV        NA
#> 154             AU100804        CMP6        1             40 eV        NA
#> 155             AU100804        CMP6        1             40 eV        NA
#> 156             AU100805        CMP7        1             50 eV        NA
#> 157             AU100805        CMP7        1             50 eV        NA
#> 158             AU100805        CMP7        1             50 eV        NA
#> 159             AU100805        CMP7        1             50 eV        NA
#> 160             AU100805        CMP7        1             50 eV        NA
#> 161             AU100805        CMP7        1             50 eV        NA
#> 162             AU100805        CMP7        1             50 eV        NA
#> 163             AU100805        CMP7        1             50 eV        NA
#> 164             AU100805        CMP7        1             50 eV        NA
#> 165             AU100805        CMP7        1             50 eV        NA
#> 166             AU100805        CMP7        1             50 eV        NA
#> 167             AU100805        CMP7        1             50 eV        NA
#> 168             AU100805        CMP7        1             50 eV        NA
#> 169             AU100805        CMP7        1             50 eV        NA
#> 170             AU100805        CMP7        1             50 eV        NA
#> 171             AU100805        CMP7        1             50 eV        NA
#> 172             AU100805        CMP7        1             50 eV        NA
#> 173             AU100805        CMP7        1             50 eV        NA
#> 174             AU100805        CMP7        1             50 eV        NA
#> 175             AU100805        CMP7        1             50 eV        NA
#> 176             AU100805        CMP7        1             50 eV        NA
#> 177             AU100805        CMP7        1             50 eV        NA
#> 178             AU100805        CMP7        1             50 eV        NA
#> 179             AU100805        CMP7        1             50 eV        NA
#>     splash instrument_type          instrument precursor_mz spectrum_id
#> 1     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 2     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 3     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 4     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 5     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 6     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 7     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 8     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 9     <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 10    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 11    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 12    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 13    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 14    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 15    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 16    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 17    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 18    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 19    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 20    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 21    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 22    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 23    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 24    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 25    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 26    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 27    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           1
#> 28    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 29    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 30    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 31    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 32    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 33    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 34    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 35    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 36    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 37    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 38    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 39    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 40    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 41    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 42    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 43    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 44    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 45    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 46    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 47    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 48    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 49    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 50    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 51    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 52    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 53    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 54    <NA>     LC-ESI-QTOF Bruker maXis Impact     285.0208           2
#> 55    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 56    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 57    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 58    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 59    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 60    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 61    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 62    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 63    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 64    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 65    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 66    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 67    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 68    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 69    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 70    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 71    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 72    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 73    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 74    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 75    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 76    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 77    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 78    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 79    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 80    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 81    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 82    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 83    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 84    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 85    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 86    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 87    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 88    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 89    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 90    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 91    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 92    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 93    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 94    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 95    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 96    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 97    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 98    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 99    <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 100   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           3
#> 101   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 102   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 103   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 104   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 105   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 106   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 107   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 108   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 109   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 110   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 111   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 112   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 113   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 114   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 115   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 116   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           4
#> 117   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 118   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 119   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 120   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 121   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 122   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 123   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 124   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 125   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 126   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 127   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 128   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 129   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 130   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 131   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 132   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 133   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           5
#> 134   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 135   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 136   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 137   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 138   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 139   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 140   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 141   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 142   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 143   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 144   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 145   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 146   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 147   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 148   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 149   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 150   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 151   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 152   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 153   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 154   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 155   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           6
#> 156   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 157   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 158   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 159   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 160   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 161   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 162   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 163   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 164   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 165   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 166   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 167   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 168   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 169   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 170   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 171   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 172   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 173   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 174   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 175   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 176   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 177   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 178   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#> 179   <NA>     LC-ESI-QTOF Bruker maXis Impact     279.0910           7
#>           mz  intensity
#> 1    53.0389   0.594951
#> 2    54.0333   0.566811
#> 3    55.0178   0.522592
#> 4    60.0552   0.542692
#> 5    65.0382   3.822962
#> 6    66.0423   0.506512
#> 7    68.0490   7.963499
#> 8    78.0333   0.727609
#> 9    79.0177   1.057244
#> 10   92.0498   7.702203
#> 11   93.0532   0.731629
#> 12   96.0443   0.623091
#> 13  108.0457  12.172375
#> 14  109.0483   1.181862
#> 15  110.0609   4.904325
#> 16  120.0562   3.095353
#> 17  130.0172   5.656054
#> 18  132.0138   1.515517
#> 19  156.0118 100.000000
#> 20  157.0150   8.884065
#> 21  158.0080   3.891301
#> 22  174.0228   0.751729
#> 23  184.0757   0.619071
#> 24  191.9647   0.590931
#> 25  219.0438   0.723589
#> 26  285.0221   3.694324
#> 27  287.0184   0.840167
#> 28   53.0389   0.594951
#> 29   54.0333   0.566811
#> 30   55.0178   0.522592
#> 31   60.0552   0.542692
#> 32   65.0382   3.822962
#> 33   66.0423   0.506512
#> 34   68.0490   7.963499
#> 35   78.0333   0.727609
#> 36   79.0177   1.057244
#> 37   92.0498   7.702203
#> 38   93.0532   0.731629
#> 39   96.0443   0.623091
#> 40  108.0457  12.172375
#> 41  109.0483   1.181862
#> 42  110.0609   4.904325
#> 43  120.0562   3.095353
#> 44  130.0172   5.656054
#> 45  132.0138   1.515517
#> 46  156.0118 100.000000
#> 47  157.0150   8.884065
#> 48  158.0080   3.891301
#> 49  174.0228   0.751729
#> 50  184.0757   0.619071
#> 51  191.9647   0.590931
#> 52  219.0438   0.723589
#> 53  285.0221   3.694324
#> 54  287.0184   0.840167
#> 55   53.0379   0.894101
#> 56   54.0335   0.661867
#> 57   55.0176   0.598003
#> 58   65.0379   8.717487
#> 59   68.0491  13.013818
#> 60   69.0329   1.640153
#> 61   78.0334   1.477589
#> 62   79.0178   2.261379
#> 63   80.0489   1.431143
#> 64   81.0444   1.950766
#> 65   82.0284   0.606712
#> 66   92.0499  30.585230
#> 67   93.0558   2.844868
#> 68   94.0647   1.686600
#> 69   95.0608   3.027752
#> 70   96.0443   1.300511
#> 71  108.0461  33.946818
#> 72  109.0497   2.360079
#> 73  110.0616   6.107757
#> 74  111.0651   0.519624
#> 75  120.0565   1.962378
#> 76  122.0716   6.078727
#> 77  123.0794   2.246865
#> 78  124.0872  71.211681
#> 79  125.0905   6.398049
#> 80  126.0663  17.911054
#> 81  127.0697   0.595100
#> 82  156.0117  82.855318
#> 83  157.0148   5.739085
#> 84  158.0072   1.544357
#> 85  174.0224   1.106015
#> 86  186.0334  11.263353
#> 87  187.0368   0.775081
#> 88  188.0128   1.637250
#> 89  188.0291   0.534138
#> 90  204.0445 100.000000
#> 91  205.0473   6.972829
#> 92  206.0406   3.358686
#> 93  213.1141  18.259405
#> 94  214.1167   2.241059
#> 95  215.0927   3.071296
#> 96  215.1291   1.320831
#> 97  279.0925  61.483976
#> 98  280.0953   8.438806
#> 99  281.0725   7.837901
#> 100 282.0742   1.222132
#> 101 122.0703   0.766124
#> 102 124.0861  36.693459
#> 103 125.0892   1.930893
#> 104 149.0227   0.828453
#> 105 156.0104  53.249536
#> 106 157.0129   2.999571
#> 107 158.0061   1.778967
#> 108 174.0209   0.627183
#> 109 186.0321  22.621444
#> 110 187.0346   1.719235
#> 111 188.0285   0.646661
#> 112 204.0431 100.000000
#> 113 213.1128   8.749399
#> 114 214.1159   1.407591
#> 115 215.1281   0.658348
#> 116 279.0909  80.894937
#> 117 108.0441   1.285794
#> 118 122.0704   6.630847
#> 119 123.0781   2.170942
#> 120 124.0861 100.000000
#> 121 125.0889   6.093546
#> 122 149.0221   1.388285
#> 123 156.0106  50.043481
#> 124 158.0064   1.615007
#> 125 186.0323  15.118951
#> 126 187.0355   1.323064
#> 127 196.0858   1.220573
#> 128 204.0429  70.964035
#> 129 205.0455   4.931983
#> 130 213.1123  22.610100
#> 131 214.1155   3.003292
#> 132 215.1283   0.804398
#> 133 279.0903   3.580968
#> 134 108.0445   1.153673
#> 135 122.0702   5.323878
#> 136 123.0772   2.202467
#> 137 124.0862 100.000000
#> 138 125.0890   6.847126
#> 139 134.0701   0.714179
#> 140 149.0224   1.747990
#> 141 154.0624   0.644259
#> 142 155.0685   0.624282
#> 143 156.0104  10.373071
#> 144 157.0126   0.933926
#> 145 172.0852   0.564351
#> 146 186.0324   3.845578
#> 147 196.0852   5.209010
#> 148 197.0903   1.378415
#> 149 198.0888   2.362283
#> 150 204.0427  15.422264
#> 151 205.0463   0.869001
#> 152 206.0375   0.759127
#> 153 212.1036   0.659242
#> 154 213.1121  18.109174
#> 155 214.1152   2.577036
#> 156 108.0453   1.770916
#> 157 122.0703   2.803951
#> 158 123.0780   2.792598
#> 159 124.0859 100.000000
#> 160 125.0891   7.901010
#> 161 149.0231   1.623340
#> 162 154.0639   2.111477
#> 163 155.0605   2.463390
#> 164 155.0714   2.690430
#> 165 156.0100   2.713134
#> 166 169.0745   1.475763
#> 167 171.0781   1.555228
#> 168 172.0869   1.271427
#> 169 181.0634   0.930866
#> 170 186.1022   1.033034
#> 171 195.0786   1.555228
#> 172 196.0859   7.628562
#> 173 197.0856   3.871041
#> 174 198.0886   5.903054
#> 175 199.0904   0.998978
#> 176 204.0438   2.622318
#> 177 212.1048   2.327165
#> 178 213.1122   9.342718
#> 179 214.1153   1.725508